% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aBKMR.R
\name{a_kmbayes}
\alias{a_kmbayes}
\title{the parameters in function a_kmbayes is same as those in kmbayes in packages "bkmr"}
\usage{
a_kmbayes(
  y,
  Z,
  X = NULL,
  iter = 1000,
  family = "gaussian",
  id = NULL,
  verbose = TRUE,
  Znew = NULL,
  starting.values = NULL,
  control.params = NULL,
  varsel = FALSE,
  groups = NULL,
  knots = NULL,
  ztest = NULL,
  rmethod = "varying",
  est.h = TRUE
)
}
\description{
the parameters in function a_kmbayes is same as those in kmbayes in packages "bkmr"
}
