% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aBKMR.R
\name{a_OverallRiskSummaries}
\alias{a_OverallRiskSummaries}
\title{used for plot}
\usage{
a_OverallRiskSummaries(
  fit,
  y = NULL,
  Z = NULL,
  X = NULL,
  qs = seq(0.25, 0.75, by = 0.05),
  q.fixed = 0.5,
  method = "approx",
  sel = NULL,
  data.comps
)
}
\arguments{
\item{data.comps}{a list including objects used for approximate gaussian process, which is generated by function "a_kmbayes"}
}
\description{
used for plot
}
