% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aBKMR.R
\name{a_SingVarRiskSummaries}
\alias{a_SingVarRiskSummaries}
\title{used for plot}
\usage{
a_SingVarRiskSummaries(
  fit,
  y = NULL,
  Z = NULL,
  X = NULL,
  data.comps,
  which.z = 1:ncol(Z),
  qs.diff = c(0.25, 0.75),
  q.fixed = c(0.25, 0.5, 0.75),
  method = "approx",
  sel = NULL,
  z.names = colnames(Z),
  ...
)
}
\arguments{
\item{data.comps}{a list including objects used for approximate gaussian process, which is generated by function "a_kmbayes"}
}
\description{
used for plot
}
