% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aBKMR.R
\name{com_2interaction}
\alias{com_2interaction}
\title{used for bivariate interaction estimation}
\usage{
com_2interaction(var1, var2, bivar_obj)
}
\arguments{
\item{var1}{first variable}

\item{var2}{second variable}

\item{bivar_obj}{the object obtained by a_PredictorResponseBivar}
}
\description{
used for bivariate interaction estimation
}
