% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aBKMR.R
\name{com_je}
\alias{com_je}
\title{used for joint effect estimation}
\usage{
com_je(postmean, postvar, X, B = 1000)
}
\arguments{
\item{postmean}{prrdicted values of y}

\item{postvar}{variance-covariance matrix of predicted y}

\item{B}{the number of bootstrap. By default, it is 1000}
}
\description{
used for joint effect estimation
}
