% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aBKMR.R
\name{com_ue}
\alias{com_ue}
\title{used for univariate effect estimation}
\usage{
com_ue(ue_obj, s)
}
\arguments{
\item{ue_obj}{data frame of univariate effect obtained by a_PredictorResponseUnivar}

\item{s}{the number of g-formula resampling}
}
\description{
used for univariate effect estimation
}
