% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_python_env.R
\name{setup_python_env}
\alias{setup_python_env}
\title{Initialize Python environment for aBKMR}
\usage{
setup_python_env(envname = "aBKMR-env")
}
\description{
Run this once in an interactive R session to create
  a conda env and install required Python packages.
}
