% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aBKMR.R
\name{sim_data}
\alias{sim_data}
\title{Simulate data}
\usage{
sim_data(
  n = 100,
  M = 5,
  sigsq.true = 0.5,
  hfun = 3,
  beta.true = 2,
  Zgen = "norm",
  ind = 1:2,
  family = "gaussian"
)
}
\arguments{
\item{n}{number of observations}

\item{M}{number of exposures}

\item{sigsq.true}{Variance of normally distributed residual error}

\item{hfun}{An integer from 1 to 3 identifying which predictor-response function to generate}

\item{beta.true}{coefficient of covariates}

\item{Zgen}{Method for generating the matrix Z of exposure variables, taking one of the values c("unif", "norm", "corr", "realistic")}

\item{ind}{select which predictor(s) will be included in the h function; how many predictors that can be included will depend on which h function is being used}

\item{family}{a description of the error distribution and link function to be used in the model. Currently implemented for gaussian and binomial families}
}
\description{
used for generating simulated data
}
\details{
advanced BKMR method

This packages achieves acceleration of BKMR and provide quantitative estimates
It includes three parts: weighted sampling, approximate gaussian process, and quantization
}
