% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aBKMR.R
\name{trend}
\alias{trend}
\title{used for multivariate interaction estimation}
\usage{
trend(beta, se)
}
\arguments{
\item{beta}{coefficient of a variable}

\item{se}{se of beta}
}
\description{
used for multivariate interaction estimation
}
