% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aBKMR.R
\name{a_OverallRiskSummaries}
\alias{a_OverallRiskSummaries}
\title{used for overall effect plot}
\usage{
a_OverallRiskSummaries(
  fit,
  y = NULL,
  Z = NULL,
  X = NULL,
  qs = seq(0.25, 0.75, by = 0.05),
  q.fixed = 0.5,
  method = "approx",
  sel = NULL,
  data.comps
)
}
\arguments{
\item{fit}{An object containing the results returned by a the kmbayes function}

\item{y}{a vector of outcome data of length n}

\item{Z}{an n-by-M matrix of predictor variables to be included in the h function. Each row represents an observation and each column represents an predictor}

\item{X}{an n-by-K matrix of covariate data where each row represents an observation and each column represents a covariate. Should not contain an intercept column}

\item{qs}{vector of quantiles at which to calculate the overall risk summary}

\item{q.fixed}{a second quantile at which to compare the estimated h function}

\item{method}{method for obtaining posterior summaries at a vector of new points. Options are "approx" and "exact"; defaults to "approx", which is faster particularly for large datasets; see details}

\item{sel}{selects which iterations of the MCMC sampler to use for inference; see details}

\item{data.comps}{a list including objects used for approximate gaussian process, which is generated by function "a_kmbayes"}
}
\description{
used for overall effect plot
}
