% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aBKMR.R
\name{a_OverallRiskSummaries_vary}
\alias{a_OverallRiskSummaries_vary}
\title{output}
\usage{
a_OverallRiskSummaries_vary(
  fit,
  y = NULL,
  Z = NULL,
  X = NULL,
  newz = NULL,
  method = "approx",
  sel = NULL,
  data.comps,
  point1 = NULL
)
}
\arguments{
\item{fit}{An object containing the results returned by a the kmbayes function}

\item{y}{a vector of outcome data of length n}

\item{Z}{an n-by-M matrix of predictor variables to be included in the h function. Each row represents an observation and each column represents an predictor}

\item{X}{an n-by-K matrix of covariate data where each row represents an observation and each column represents a covariate. Should not contain an intercept column}

\item{newz}{A dataset of counterfactual scenarios of interest, representing hypothetical exposure profiles. By default, all exposures are increased simultaneously from their 25% quantiles to 75% quantiles with 1% quantiles.}

\item{method}{method for obtaining posterior summaries at a vector of new points. Options are "approx" and "exact"; defaults to "approx", which is faster particularly for large datasets; see details}

\item{sel}{selects which iterations of the MCMC sampler to use for inference; see details}

\item{data.comps}{a list including objects used for approximate gaussian process, which is generated by function "a_kmbayes"}

\item{point1}{reference point. By default, it is specified as c(median(X1), median(X2), median(X3), …). You can set it to any value of interest – for example, the standard (recommended) exposure level.}

\item{quants}{vector of quantiles at which to calculate the overall risk summary}
}
\description{
used for generating plot and estimate of effect
used for overall effect plot with varying values for exposures
}
