% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aBKMR.R
\name{a_PredictorResponseBivar}
\alias{a_PredictorResponseBivar}
\title{used for plot}
\usage{
a_PredictorResponseBivar(
  fit,
  y = NULL,
  Z = NULL,
  X = NULL,
  data.comps = data.comps,
  z.pairs = NULL,
  method = "approx",
  ngrid = 50,
  q.fixed = 0.5,
  sel = NULL,
  min.plot.dist = 0.5,
  center = TRUE,
  z.names = colnames(Z),
  verbose = TRUE,
  ...
)
}
\arguments{
\item{data.comps}{a list including objects used for approximate gaussian process, which is generated by function "a_kmbayes"}
}
\description{
used for plot
}
