% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aBKMR.R
\name{a_PredictorResponseUnivar}
\alias{a_PredictorResponseUnivar}
\title{used for plot}
\usage{
a_PredictorResponseUnivar(
  fit,
  y = NULL,
  Z = NULL,
  X = NULL,
  data.comps,
  which.z = 1:ncol(Z),
  method = "approx",
  ngrid = 50,
  q.fixed = 0.5,
  sel = NULL,
  min.plot.dist = Inf,
  center = TRUE,
  z.names = colnames(Z),
  ...
)
}
\arguments{
\item{data.comps}{a list including objects used for approximate gaussian process, which is generated by function "a_kmbayes"}
}
\description{
used for plot
}
