% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aBKMR.R
\name{a_PredictorResponseUnivar}
\alias{a_PredictorResponseUnivar}
\title{used for univariate concentration-response curves of all variables}
\usage{
a_PredictorResponseUnivar(
  fit,
  y = NULL,
  Z = NULL,
  X = NULL,
  data.comps,
  which.z = 1:ncol(Z),
  method = "approx",
  ngrid = 50,
  q.fixed = 0.5,
  sel = NULL,
  min.plot.dist = Inf,
  center = TRUE,
  z.names = colnames(Z),
  ...
)
}
\arguments{
\item{fit}{An object containing the results returned by a the kmbayes function}

\item{y}{a vector of outcome data of length n}

\item{Z}{an n-by-M matrix of predictor variables to be included in the h function. Each row represents an observation and each column represents an predictor}

\item{X}{an n-by-K matrix of covariate data where each row represents an observation and each column represents a covariate. Should not contain an intercept column}

\item{data.comps}{a list including objects used for approximate gaussian process, which is generated by function "a_kmbayes"}

\item{which.z}{vector indicating which variables (columns of Z) for which the summary should be computed}

\item{method}{method for obtaining posterior summaries at a vector of new points. Options are "approx" and "exact"; defaults to "approx", which is faster particularly for large datasets; see details}

\item{ngrid}{number of grid points to cover the range of each predictor (column in Z)}

\item{q.fixed}{vector of quantiles at which to fix the remaining predictors in Z}

\item{sel}{logical expression indicating samples to keep; defaults to keeping the second half of all samples}

\item{min.plot.dist}{specifies a minimum distance that a new grid point needs to be from an observed data point in order to compute the prediction; points further than this will not be computed}

\item{center}{flag for whether to scale the exposure-response function to have mean zero}

\item{z.names}{optional vector of names for the columns of z}

\item{...}{other arguments to pass on to the prediction function}
}
\description{
used for univariate concentration-response curves of all variables
}
