% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aBKMR.R
\name{a_PredictorResponseUnivar}
\alias{a_PredictorResponseUnivar}
\title{used for univariate effect plot}
\usage{
a_PredictorResponseUnivar(
  fit,
  y = NULL,
  Z = NULL,
  X = NULL,
  quants = seq(0.25, 0.75, 0.1),
  method = "approx",
  sel = NULL,
  data.comps,
  point1 = NULL
)
}
\arguments{
\item{fit}{An object containing the results returned by a the kmbayes function}

\item{y}{a vector of outcome data of length n}

\item{Z}{an n-by-M matrix of predictor variables to be included in the h function. Each row represents an observation and each column represents an predictor}

\item{X}{an n-by-K matrix of covariate data where each row represents an observation and each column represents a covariate. Should not contain an intercept column}

\item{quants}{specify the quantiles of exposures}

\item{method}{method for obtaining posterior summaries at a vector of new points. Options are "approx" and "exact"; defaults to "approx", which is faster particularly for large datasets; see details}

\item{sel}{logical expression indicating samples to keep; defaults to keeping the second half of all samples}

\item{data.comps}{a list including objects used for approximate gaussian process, which is generated by function "a_kmbayes"}
}
\description{
used for univariate effect plot
}
