% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aBKMR.R
\name{a_SingVarRiskSummaries}
\alias{a_SingVarRiskSummaries}
\title{used for multivariate interaction plot}
\usage{
a_SingVarRiskSummaries(
  fit,
  y = NULL,
  Z = NULL,
  X = NULL,
  data.comps,
  which.z = 1:ncol(Z),
  qs.diff = c(0.25, 0.75),
  q.fixed = c(0.25, 0.5, 0.75),
  method = "approx",
  sel = NULL,
  z.names = colnames(Z),
  ...
)
}
\arguments{
\item{fit}{An object containing the results returned by a the kmbayes function}

\item{y}{a vector of outcome data of length n}

\item{Z}{an n-by-M matrix of predictor variables to be included in the h function. Each row represents an observation and each column represents an predictor}

\item{X}{an n-by-K matrix of covariate data where each row represents an observation and each column represents a covariate. Should not contain an intercept column}

\item{data.comps}{a list including objects used for approximate gaussian process, which is generated by function "a_kmbayes"}

\item{which.z}{vector indicating which variables (columns of Z) for which the summary should be computed}

\item{qs.diff}{vector indicating the two quantiles at which to compute the single-predictor risk summary}

\item{q.fixed}{vector indicating the two quantiles at which to fix all of the remaining exposures in Z}

\item{method}{method for obtaining posterior summaries at a vector of new points. Options are "approx" and "exact"; defaults to "approx", which is faster particularly for large datasets; see details}

\item{sel}{logical expression indicating samples to keep; defaults to keeping the second half of all samples}

\item{z.names}{optional vector of names for the columns of z}

\item{...}{other arguments to pass on to the prediction function}
}
\description{
used for multivariate interaction plot
}
