% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/_PACKAGE.R
\docType{package}
\name{aBKMR-package}
\alias{aBKMR}
\alias{aBKMR-package}
\title{aBKMR: Advanced Bayesian Kernel Machine Regression for Large-Scale Exposome Studies}
\description{
Provides an implementation of Advanced Bayesian Kernel Machine Regression for large-scale exposome data. It dramatically reduces both computation time and memory usage. The package also offers functions for visualizing univariate, bivariate, and multivariate exposure response surfaces in R..
}
\keyword{internal}
