% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aBKMR.R
\name{com_je}
\alias{com_je}
\title{used for joint effect estimation}
\usage{
com_je(je_df, s)
}
\arguments{
\item{je_df}{data frame of overall effect obtained by a_OverallRiskSummaries}

\item{s}{the number of g-formula resampling}
}
\description{
used for joint effect estimation
}
