% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_code.R
\name{makeKpart}
\alias{makeKpart}
\title{There are some internal functions}
\usage{
makeKpart(r, Z1, Z2 = NULL)
}
\arguments{
\item{y}{outcome}

\item{x}{covariates}

\item{Z}{exposures}
}
\value{
objects for consquent computation
}
\description{
There are some internal functions
}
\keyword{internal}
