% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aBKMR.R
\name{sam_py_r}
\alias{sam_py_r}
\title{used for knot sampling}
\usage{
sam_py_r(R, nd, num_nn, P = -20, Q = 20, max_loop = 20, w = F)
}
\arguments{
\item{R}{exposures}

\item{nd}{number of the representative observations}

\item{num_nn}{number of nearest neighbor}

\item{P}{computation parameter}

\item{Q}{computation parameter}

\item{max_loop}{max number of loop finding knots}

\item{w}{whether the weight is used}
}
\description{
used for knot sampling
}
